fix: compile errors

+++ src/scene.cpp
@@ -1,5 +1,6 @@
 #include "scene.h"
 #include <numbers>
+#include <cmath>
 #include <SDL3/SDL_rect.h>
 
 static float time = 0.0f;
@@ -11,8 +12,8 @@ void Scene::draw(SDL_Renderer& renderer, float delta_time) const {
     SDL_FRect rect;
 
     SDL_SetRenderDrawColor(&renderer, 125, 200, 255, SDL_ALPHA_OPAQUE);  /* blue, full alpha */
-    rect.x = 100.0f + 200.0f * cosf(time * 2.0f * PI * rpm / 60.0f);
-    rect.y = 100.0f + 200.0f * sinf(time * 2.0f * PI * rpm / 60.0f);
+    rect.x = 100.0f + 200.0f * std::cos(time * 2.0f * PI * rpm / 60.0f);
+    rect.y = 100.0f + 200.0f * std::sin(time * 2.0f * PI * rpm / 60.0f);
     rect.w = 440.0f;
     rect.h = 280.0f;
     SDL_RenderFillRect(&renderer, &rect);
